PHASE_1/API_SourceCode/requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
cycler==0.11.0
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
filelock==3.6.0
fonttools==4.30.0
h11==0.13.0
httptools==0.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
kiwisolver==1.4.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
orjson==3.6.7
packaging==21.3
parsel==1.6.0
Pillow==9.0.1
pluggy==1.0.0
Protego==0.2.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic!=1.8,!=1.8.1,<1.9.0,>=1.7.4
PyDispatcher==2.0.5
pymongo==4.0.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
spacy
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tldextract==3.2.0
tomli==2.0.1
Twisted==22.2.0
typing-extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
w3lib==1.22.0
watchgod==0.8
websockets==10.2
zope.interface==5.4.0
